Fixing inconsistency
Since we don't want to have the CID added to the DTLSCompressed structure we cannot reference it here in this MAC function. For this reason, I am changing the text to refer to the message fields directly. I hope this helps. It is a kind-of artificial problem since we are not recommending to use the DTLS compression feature anymore anyway.